# Conflicts:
#	Binaries/Win64/UnrealEditor-CouchGame.dll
#	Content/_project/BP/BP_ThirdPersonCharacterTest.uasset
#	Plugins/CameraSystemPlugin/Binaries/Win64/UnrealEditor-CameraSystemPlugin.dll

+++ Source/CouchGame/Public/Utils/ArrayUtils.inl
@@ -4,6 +4,6 @@ template <typename T>
 void UArrayUtils::GetRandomElement(const TArray<T>& Array, T& outElement)
 {
 	if (Array.IsEmpty()) return;
-	int32 RandRange = FMath::RandRange(0, Array.Num() - 1);
+	uint16 RandRange = FMath::RandRange(0, Array.Num() - 1);
 	outElement = Array[RandRange];
 }
